feat: enhance user preferences and navbar components, add favorite team and league schemas
