+++ client/src/components/Navbar/navbar.module.scss
@@ -1,70 +1,72 @@
 @use '../../theme/theme' as *;
+
 $bg-color: #f9f9f9;
 $text-color: #333;
 $link-hover-color: #000;
 
 .navbar {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    background-color: $background;
-    padding: 0.5rem 1rem;
-    border-bottom: 1px solid #ddd;
-    font-family: Arial, sans-serif;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background-color: $background;
+  padding: 0.5rem 1rem;
+  border-bottom: 1px solid #ddd;
+  font-family: Arial, sans-serif;
 }
 
 .brand {
-    display: flex;
-    align-items: center;
-    cursor: pointer;
+  display: flex;
+  align-items: center;
+  cursor: pointer;
 
-    .icon {
-        font-size: 1.25rem;
-        margin-right: 0.5rem;
-    }
-    gap: 0.5rem;
+  .icon {
+    font-size: 1.25rem;
+    margin-right: 0.5rem;
+  }
 
-    & > svg {
-      width: 1.2rem;
-      height: 1.2rem;
-    }
+  gap: 0.5rem;
 
-    .title {
-        font-size: 1.25rem;
-        font-weight: bold;
-        color: $text-color;
-    }
+  & > svg {
+    width: 1.2rem;
+    height: 1.2rem;
+  }
+
+  .title {
+    font-size: 1.25rem;
+    font-weight: bold;
+    color: $text-color;
+  }
 }
 
 .navLinks {
-    display: flex;
-    gap: 1.5rem;
+  display: flex;
+  gap: 1.5rem;
 
-    a {
-        text-decoration: none;
-        color: $text-color;
-        font-weight: 700;
+  a {
+    text-decoration: none;
+    color: $text-color;
+    font-weight: 700;
 
-        &:hover {
-            color: $link-hover-color;
-            text-decoration: underline;
-            cursor: pointer;
-        }
+    &:hover {
+      color: $link-hover-color;
+      text-decoration: underline;
+      cursor: pointer;
     }
+  }
 }
 
 .userProfile {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
 
-    .avatar {
-        width: 40px;
-        height: 40px;
-        border-radius: 50%;
-        object-fit: cover;
-        cursor: pointer;
-    }
+  .avatar {
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    object-fit: cover;
+    cursor: pointer;
+  }
 }
 
 .headerLeft {
@@ -80,3 +82,10 @@ $link-hover-color: #000;
   gap: 2rem;
 }
 
+.preferencesModal {
+  :global {
+    .ant-modal-content {
+      width: 44rem;
+    }
+  }
+}
