+++ client/src/components/NavbarUserDropdown/NavbarUserDropdown.tsx
@@ -5,72 +5,74 @@ import { getPictureSrcUrl } from '@/utils/picture.utils';
 import { EditOutlined, LogoutOutlined } from '@ant-design/icons';
 
 export interface NavbarUserDropdownProps {
-  showModal: () => void;
-  showPreferencesModal: () => void;
+    showModal: () => void;
+    showPreferencesModal: () => void;
 }
 
 const { Text } = Typography;
 
 export const NavbarUserDropdown = ({ showModal, showPreferencesModal }: NavbarUserDropdownProps) => {
-  const { loggedInUser, logout } = useAuth();
+    const { loggedInUser, logout } = useAuth();
 
-  const items: MenuProps['items'] = [
-    {
-      key: 'profile',
-      label: loggedInUser ? (
-        <div className={classes.dropdownContainer}>
-          <div className={classes.profileImageContainer}>
-            <img
-              src={getPictureSrcUrl(loggedInUser?.picture)}
-              alt="User Avatar"
-              className={classes.profileImage}
-            />
-          </div>
+    const items: MenuProps['items'] = [
+        {
+            key: 'profile',
+            label: loggedInUser ? (
+                <div className={classes.dropdownContainer}>
+                    <div className={classes.profileImageContainer}>
+                        <img
+                            src={getPictureSrcUrl(loggedInUser?.picture)}
+                            alt="User Avatar"
+                            className={classes.profileImage}
+                        />
+                    </div>
 
-          <div className={classes.textCenter}>
-            <Text className={classes.profileName}>{loggedInUser.username}</Text>
-            <Text className={classes.profileEmail}>{loggedInUser.email}</Text>
-          </div>
+                    <div className={classes.textCenter}>
+                        <Text className={classes.profileName}>{loggedInUser.username}</Text>
+                        <Text className={classes.profileEmail}>{loggedInUser.email}</Text>
+                    </div>
 
-          <Button
-            type="link"
-            icon={<EditOutlined />}
-            className={classes.customizeProfileButton}
-            onClick={showModal}
-          >
-            Edit User
-          </Button>
+                    <div className={classes.textCenter}>
+                        <Button
+                            type="link"
+                            icon={<EditOutlined />}
+                            className={classes.customizeProfileButton}
+                            onClick={showModal}
+                        >
+                            Edit User
+                        </Button>
 
-          <Button
-            type="link"
-            icon={<EditOutlined />}
-            className={classes.customizeProfileButton}
-            onClick={showPreferencesModal}
-          >
-            Edit Preferences
-          </Button>
+                        <Button
+                            type="link"
+                            icon={<EditOutlined />}
+                            className={classes.customizeProfileButton}
+                            onClick={showPreferencesModal}
+                        >
+                            Edit Preferences
+                        </Button>
+                    </div>
 
-          <Divider className={classes.divider} />
+                    <Divider className={classes.divider} />
 
-          <Button
-            type="primary"
-            onClick={logout}
-            danger
-            icon={<LogoutOutlined />}
-            className={classes.logoutButton}
-          >
-            Log Out
-          </Button>
-        </div>
-      ) : null,
-    },
-  ];
+                    <Button
+                        type="primary"
+                        onClick={logout}
+                        danger
+                        icon={<LogoutOutlined />}
+                        className={classes.logoutButton}
+                    >
+                        Log Out
+                    </Button>
+                </div>
+            ) : null,
+        },
+    ];
 
-  return (
-    <Dropdown menu={{ items }} trigger={['click']} placement="bottomRight">
-      {loggedInUser && (
-        <img src={getPictureSrcUrl(loggedInUser.picture)} alt="User Avatar" className={classes.avatarSmall} />
-      )}
-    </Dropdown>
-  );
+    return (
+        <Dropdown menu={{ items }} trigger={['click']} placement="bottomRight">
+            {loggedInUser && (
+                <img src={getPictureSrcUrl(loggedInUser.picture)} alt="User Avatar" className={classes.avatarSmall} />
+            )}
+        </Dropdown>
+    );
 };
