+++ client/src/pages/UserPreferences/preferences-body-model.scss
@@ -2,15 +2,16 @@
   max-width: 600px;
   margin: 0 auto;
   padding: 24px;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
 
   .intro {
-    margin-bottom: 24px;
     font-size: 16px;
     color: #444;
   }
 
   .form-group {
-    margin-bottom: 24px;
 
     label {
       display: block;
@@ -24,6 +25,7 @@
         font-size: 14px;
       }
     }
+
   }
 
   .save-btn {
@@ -47,4 +49,9 @@
   justify-content: center;
   align-items: center;
   padding: 100px 0;
+}
+
+
+.ant-select {
+  width: 100%;
 }
