+++ client/src/pages/UserPreferences/preferences-body.const.ts
@@ -1,0 +1,1 @@
+export const MAX_ITEMS_PER_SELECT = 3;
