+++ server/src/models/user.model.ts
@@ -1,6 +1,21 @@
 import { z } from 'zod';
 import { ObjectIdToString, zodDate } from '../utils/zod.utils';
-import { CityInfoSchema, TeamNoLogoSchema } from './packages/package.model';
+import { CityInfoSchema } from './packages/package.model';
+
+export const FavoriteTeamSchema = z.object({
+    name: z.string().describe('Team name'),
+    id: z.number().describe('Unique identifier of the team'),
+    logo: z.string().describe('URL of the team logo'),
+});
+
+export const FavoriteLeagueSchema = z.object({
+    name: z.string().describe('League name'),
+    id: z.number().describe('Unique identifier of the league'),
+    logo: z.string().describe('URL of the league logo'),
+});
+
+export type FavoriteTeam = z.infer<typeof FavoriteTeamSchema>;
+export type FavoriteLeague = z.infer<typeof FavoriteLeagueSchema>;
 
 export const UserSchema = z.object({
     username: z.string(),
@@ -11,9 +26,9 @@ export const UserSchema = z.object({
     createdAt: zodDate,
     updatedAt: zodDate,
     refreshTokens: z.string().array().optional(),
-    favoriteTeams: z.array(TeamNoLogoSchema).optional(),
+    favoriteTeams: FavoriteTeamSchema.array().optional(),
+    favoriteLeagues: FavoriteLeagueSchema.array().optional(),
     homeAirport: CityInfoSchema.optional(),
-    favoriteLeagues: z.string().array().optional(),
     isFirstVisit: z.boolean().optional(),
 });
